Add files via upload
The data deals with the Superstore data of the United States.  The data overview is built on the Power BI dashboard. Basically, it is 3 Insight 

Insight 1 - The first dashboard deals with overall sales and profit in the United States region 

Insight 2 - The Second dashboard deals with the Overall sales and profit with each region in  the united states

Insight 3 - The  third dashboard deals with the category analysis and the sales and profit of each category and subcategory